03_advanced_level/02_flask_blog_app/api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.4.0